Update widget.js
ABSOLUTELY HUGE MIND-BLOWING CHANGE: basically the widget would not change the width of the container when it was transitioning from a song with a long name to another song with a long name, this was a complete oversight in the code where I put 'container.width - 70' instead of just doing 260-70 and then putting the number there, this means that when the long song it playing, the calculation for the next song, is checking if the song name was longer than lets say 500-70, which if it wasn't, then it would default back to 260px. anyway, big oversight, now its just a number and it works perfectly!